Soluciono el envio de GameObjectDTO
No se estaba iniciando el hilo receiver del client y ademas se estaba mandando mal la cantidad de elementos dentro de GameObjectDTO

+++ src/server/server_serializer.cpp
@@ -94,8 +94,7 @@ void ServerSerializer::send_game(GameDTO game_dto, bool *was_closed)
 {
     std::vector<char> buffer;
     std::vector<GameObjectDTO> objects_game = game_dto.get_objects_game();
-    buffer.push_back(objects_game.size());
-
+    buffer.push_back(uint8_t(objects_game.size()));
     for (const auto &obj : objects_game)
     {
         buffer.push_back(obj.state);
